pyproject.toml
[tool.poetry]
name = "sponsorbook"
version = "0.1.0"
description = ""
authors = ["edsinkevic <sinkevicedwin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.3.1"
pymongo = "^4.3.3"
fastapi = "^0.95.1"
uvicorn = "^0.21.1"
motor = "^3.1.2"
motor-types = "^1.0.0b1"
pytest-asyncio = "^0.21.0"
python-multipart = "^0.0.6"
argon2-cffi = "^21.3.0"
pyjwt = "^2.7.0"
structlog = "^23.1.0"
python-json-logger = "^2.0.7"
pandas = "^2.0.3"
openpyxl = "^3.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
